+++ src/main/kotlin/com/aoc/intcode/amplifier/Amplifier.kt
@@ -1,7 +1,7 @@
 package com.aoc.intcode.amplifier
 
 interface Amplifier {
-    fun inputSignal(inputSignal: Int)
+    fun inputSignal(inputSignal: Long)
     fun outputsTo(amplifier: Amplifier)
     fun loadAmplifierControllerSoftware(software: String)
 }
