+++ src/main/kotlin/com/aoc/intcode/amplifier/series/InitialAmplifier.kt
@@ -2,7 +2,7 @@ package com.aoc.intcode.amplifier.series
 
 import com.aoc.intcode.amplifier.BaseAmplifier
 
-class InitialAmplifier(phaseSetting: Int) : BaseAmplifier(phaseSetting) {
+class InitialAmplifier(phaseSetting: Long) : BaseAmplifier(phaseSetting) {
     fun start() {
         super.inputSignal(0)
     }
